Split `development-plugin.php` into classes
